MicroServices/AuthenticationService/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
blinker==1.8.2
click==8.1.7
dnspython==2.7.0
fastapi==0.115.4
Flask==3.0.3
Flask-JWT-Extended==4.6.0
Flask-Mail==0.10.0
Flask-PyMongo==2.3.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
passlib==1.7.4
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
pymongo==4.10.1
python-dotenv==1.0.1
sniffio==1.3.1
starlette==0.41.2
typing_extensions==4.12.2
Werkzeug==3.1.2
